Updating fields in DB using Python
In general, we try to use the smallest data type that can correctly store and represent our data. Smaller data types are usually faster, because they use less space on the disk, in memory, and in the CPU cache. They also generally require fewer CPU cycles to process.

In this Python script, we've achieved the task of assigning a job code to the various types of jobs and updated the same in the DB as well.
Similarly, we can write similar functions to do the same for other fields as well.